+++ src/modules/index.ts
@@ -1,0 +1,2 @@
+export { default as ZeroBywDownloader } from "./zerobyw";
+export { default as CopymangaDownloader } from "./copymanga";